+++ src/overtone/studio/inst.clj
@@ -16,6 +16,7 @@
         [overtone.helpers.lib]
         [overtone.libs.event])
   (:require [clojure.pprint]
+            [lentes.core :as l]
             [overtone.sc.protocols :as protocols]
             [overtone.sc.util]
             [overtone.sc.machinery.server.comms :refer [with-server-sync]]))
@@ -24,8 +25,7 @@
   (do
     (defrecord-ifn Inst [name full-name params args sdef
                          group instance-group fx-group
-                         mixer bus fx-chain
-                         volume pan
+                         mixer mixer-params volume pan bus
                          n-chans]
       (fn [this & args]
         (apply synth-player sdef params this [:tail instance-group] args))
@@ -64,12 +64,40 @@
             sndr (select 1 snd)]
         (out out-bus (balance2 sndl sndr pan volume))))))
 
-(defn inst-mixer
+(defn default-get-inst-mixer
   "Instantiate a mono or stereo inst-mixer synth."
-  [n-chans & args]
+  [{:keys [n-chans] :as _inst}]
   (if (> n-chans 1)
-    (apply stereo-inst-mixer args)
-    (apply mono-inst-mixer args)))
+    stereo-inst-mixer
+    mono-inst-mixer))
+
+(defn inst-mixer
+  "Instantiate an instrument mixer."
+  [inst & args]
+  (let [get-mixer (get @studio* ::get-inst-mixer default-get-inst-mixer)
+        mixer-synth (get-mixer inst)]
+    (apply mixer-synth args)))
+
+(defn replace-inst-mixer!
+  "Replace the mixer synth in an instrument."
+  [{:keys [n-chans mixer bus mixer-params] :as inst} new-get-inst-mixer & {:as params}]
+  (ensure-node-active! inst)
+  (let [mixer-synth (new-get-inst-mixer inst)
+        synth-params (->> (merge @mixer-params params)
+                          ;; Flatten to :key1 val1 :key2 val2 ...
+                          (mapcat identity))
+        new-mixer (apply mixer-synth
+                         [:replace @mixer]
+                         :in-bus bus
+                         synth-params)]
+    (reset! mixer new-mixer)))
+
+(defn replace-all-inst-mixer!
+  "Replace the mixer synth in all current and future instruments."
+  [new-get-inst-mixer & params]
+  (swap! studio* assoc ::get-inst-mixer new-get-inst-mixer)
+  (doseq [[_name inst] (:instruments @studio*)]
+    (apply replace-inst-mixer! inst new-get-inst-mixer params)))
 
 (defn inst-channels
   "Internal fn used for multimethod dispatch on Insts."
@@ -81,15 +109,22 @@
   "Control the volume of a single instrument."
   [inst vol]
   (ensure-node-active! inst)
-  (ctl (:mixer inst) :volume vol)
+  (ctl @(:mixer inst) :volume vol)
   (reset! (:volume inst) vol))
 
 (defn inst-pan!
   "Control the pan setting of a single instrument."
   [inst pan]
   (ensure-node-active! inst)
-  (ctl (:mixer inst) :pan pan)
+  (ctl @(:mixer inst) :pan pan)
   (reset! (:pan inst) pan))
+
+(defn inst-mixer-ctl!
+  "Control a named parameters of the output mixer of a single instrument."
+  [inst & {:as args}]
+  (ensure-node-active! inst)
+  (apply ctl @(:mixer inst) (-> args seq flatten))
+  (swap! (:mixer-params inst) merge args))
 
 (defmulti inst-fx!
   "Append an effect to an instrument channel. Returns a SynthNode or a
@@ -179,22 +214,27 @@
                                 "whilst creating an inst fx group"))
 
          imixer#    (or (:mixer new-inst#)
-                        (inst-mixer n-chans#
-                                    [:tail container-group#]
-                                    :in-bus inst-bus#))
+                        (atom ::uninitialized-mixer))
          sdef#      (synthdef sname# params# ugens# constants#)
          arg-names# (map :name params#)
          params-with-vals# (map #(assoc % :value (control-proxy-value-atom full-name# %)) params#)
-         fx-chain#  []
-         volume#    (atom DEFAULT-VOLUME)
-         pan#       (atom DEFAULT-PAN)
+         mixer-params# (atom {:volume DEFAULT-VOLUME
+                              :pan    DEFAULT-PAN})
+         volume#    (l/derive (l/key :volume) mixer-params#)
+         pan#       (l/derive (l/key :pan) mixer-params#)
          inst#      (with-meta
                       (->Inst sname# full-name# params-with-vals# arg-names# sdef#
                               container-group# instance-group# fx-group#
-                              imixer# inst-bus# fx-chain#
-                              volume# pan#
+                              imixer# mixer-params# volume# pan# inst-bus#
                               n-chans#)
                       {:overtone.helpers.lib/to-string #(str (name (:type %)) ":" (:name %))})]
+     (when (= ::uninitialized-mixer @imixer#)
+       ;; Briefly delayed mixer creation to support passing the realized inst# to
+       ;; the mixer selection function.
+       (reset! imixer#
+               (inst-mixer inst#
+                           [:tail container-group#]
+                           :in-bus inst-bus#)))
      (load-synthdef sdef#)
      (add-instrument inst#)
      (event :new-inst :inst inst#)
@@ -275,8 +315,8 @@
     (inst-pan! bar 1 -1)   ;ch1 right, ch2 left.
     (inst-volume! bar 0 1) ;mute ch1.
 
-  * Each instrument has an fx-chain to which you can add any number of
-    'fx synths' using the inst-fx function.
+  * Each instrument has an fx-group to which you can add any number of
+    'fx synths' using the inst-fx! function.
   "
   {:arglists '([name doc-string? params ugen-form])}
   [i-name & inst-form]
@@ -306,7 +346,7 @@
     (doseq [sub-node [(:fx-group inst)
                       (:group inst)
                       (:instance-group inst)
-                      (:mixer inst)]]
+                      @(:mixer inst)]]
       (node-block-until-ready sub-node))))
 
 (defn- inst-status*
@@ -314,7 +354,7 @@
   (let [sub-nodes [(:fx-group inst)
                    (:group inst)
                    (:instance-group inst)
-                   (:mixer inst)]]
+                   @(:mixer inst)]]
     (cond
      (some #(= :loading @(:status %)) sub-nodes) :loading
      (some #(= :destroyed @(:status %)) sub-nodes) :destroyed